chore: displayField model
