+++ packages/core/client/src/flow/models/fields/DisplayIconFieldModel.tsx
@@ -9,10 +9,10 @@
 
 import React from 'react';
 import { DisplayItemModel } from '@nocobase/flow-engine';
-import { ReadPrettyFieldModel } from './ReadPrettyFieldModel';
-import { Icon } from '../../../../icon';
+import { Icon } from '../../../icon';
+import { InteractiveDisplayFieldModel } from './InteractiveDisplayFieldModel';
 
-export class IconReadPrettyFieldModel extends ReadPrettyFieldModel {
+export class DisplayIconFieldModel extends InteractiveDisplayFieldModel {
   public static readonly supportedFieldInterfaces = ['icon'];
   public render() {
     const { value } = this.props;
@@ -20,6 +20,6 @@ export class IconReadPrettyFieldModel extends ReadPrettyFieldModel {
   }
 }
 
-DisplayItemModel.bindModelToInterface('IconReadPrettyFieldModel', ['icon'], {
+DisplayItemModel.bindModelToInterface('DisplayIconFieldModel', ['icon'], {
   isDefault: true,
 });
