+++ packages/core/client/src/flow/models/fields/DisplayPercentFieldModel.tsx
@@ -11,7 +11,7 @@ import React from 'react';
 import * as math from 'mathjs';
 import { DisplayItemModel } from '@nocobase/flow-engine';
 import { isNum } from '@formily/shared';
-import { ReadPrettyFieldModel } from './ReadPrettyFieldModel';
+import { InteractiveDisplayFieldModel } from './InteractiveDisplayFieldModel';
 
 const isNumberLike = (index: any): index is number => isNum(index) || /^-?\d+(\.\d+)?$/.test(index);
 
@@ -21,7 +21,7 @@ const toValue = (value: any, callback: (v: number) => number) => {
   }
   return null;
 };
-export class PercentReadPrettyFieldModel extends ReadPrettyFieldModel {
+export class DisplayPercentFieldModel extends InteractiveDisplayFieldModel {
   public static readonly supportedFieldInterfaces = ['percent'];
   // @reactive
   public render() {
@@ -38,6 +38,6 @@ export class PercentReadPrettyFieldModel extends ReadPrettyFieldModel {
   }
 }
 
-DisplayItemModel.bindModelToInterface('PercentReadPrettyFieldModel', ['percent'], {
+DisplayItemModel.bindModelToInterface('DisplayPercentFieldModel', ['percent'], {
   isDefault: true,
 });
