+++ packages/core/client/src/flow/models/fields/DisplayTextFieldModel.tsx
@@ -9,9 +9,9 @@
 
 import { DisplayItemModel, escapeT } from '@nocobase/flow-engine';
 import React from 'react';
-import { ReadPrettyFieldModel } from './ReadPrettyFieldModel';
+import { InteractiveDisplayFieldModel } from './InteractiveDisplayFieldModel';
 
-export class InputReadPrettyFieldModel extends ReadPrettyFieldModel {
+export class DisplayTextFieldModel extends InteractiveDisplayFieldModel {
   static supportedFieldInterfaces = [
     'input',
     'email',
@@ -36,11 +36,11 @@ export class InputReadPrettyFieldModel extends ReadPrettyFieldModel {
     );
   }
 }
-InputReadPrettyFieldModel.define({
+DisplayTextFieldModel.define({
   label: escapeT('Text'),
 });
 DisplayItemModel.bindModelToInterface(
-  'InputReadPrettyFieldModel',
+  'DisplayTextFieldModel',
   ['input', 'email', 'phone', 'uuid', 'attachmentURL', 'textarea', 'time', 'nanoid'],
   {
     isDefault: true,
