--- packages/core/client/src/flow/models/fields/ReadPrettyField/index.ts
@@ -1,24 +1,0 @@
-/**
- * This file is part of the NocoBase (R) project.
- * Copyright (c) 2020-2024 NocoBase Co., Ltd.
- * Authors: NocoBase Team.
- *
- * This project is dual-licensed under AGPL-3.0 and NocoBase Commercial License.
- * For more information, please refer to: https://www.nocobase.com/agreement.
- */
-
-export * from './CheckboxReadPrettyFieldModel';
-export * from './DateTimeReadPrettyFieldModel';
-export * from './InputReadPrettyFieldModel';
-export * from './NumberReadPrettyFieldModel';
-export * from './PercentReadPrettyFieldModel';
-export * from './ReadPrettyFieldModel';
-export * from './RichTextReadPrettyFieldModel';
-export * from './SelectReadPrettyFieldModel';
-export * from './ColorReadPrettyFieldModel';
-export * from './IconReadPrettyFieldModel';
-export * from './JsonReadPrettyFieldModel';
-export * from './AssociationFieldModel';
-export * from './MarkdownReadPrettyFieldModel';
-export * from './URLReadPrettyFieldModel';
-export * from './PasswordReadPrettyFieldModel';
