+++ packages/plugins/@nocobase/plugin-file-manager/src/client/index.tsx
@@ -18,7 +18,7 @@ import { AttachmentFieldInterface } from './interfaces/attachment';
 import { NAMESPACE } from './locale';
 import { storageTypes } from './schemas/storageTypes';
 import { FileCollectionTemplate } from './templates';
-import { PreviewReadPrettyFieldModel } from './fieldModels/PreviewReadPrettyFieldModel';
+import { DisplayPreviewFieldModel } from './fieldModels/DisplayPreviewFieldModel';
 
 export class PluginFileManagerClient extends Plugin {
   // refer by plugin-field-attachment-url
@@ -67,7 +67,7 @@ export class PluginFileManagerClient extends Plugin {
       FileSizeField,
     });
 
-    this.flowEngine.registerModels({ PreviewReadPrettyFieldModel });
+    this.flowEngine.registerModels({ DisplayPreviewFieldModel });
   }
 
   registerStorageType(name: string, options) {
